tidy: Explicitly convert boolean to integer for some library calls.
The clang-tidy "implicit boolean conversion" check pointed out a
couple of places where a system library is expecting a boolean value,
but the API states that the value is an integer.  These true/false
values could be converted to 1/0, but the intent seems more obvious if
the true/false values are kept and are explicitly converted to integer
values.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html

+++ mythplugins/mytharchive/mytharchivehelper/main.cpp
@@ -1946,7 +1946,7 @@ static int getFileInfo(QString inFile, QString outFile, int lenMethod)
 
         buf[0]=0;
         if (avctx)
-            avcodec_string(buf, sizeof(buf), avctx, false);
+            avcodec_string(buf, sizeof(buf), avctx, static_cast<int>(false));
 
         switch (st->codecpar->codec_type)
         {
